[ YC-632 ] YUM webapp
- added force no filter feature
- extracted some UI configs into globals
- refactored category table
- added new attributes table
- added custom date formatting pipes (ng2 OOTB broken on various browsers)
- added case insensitive searching at DB level for categories and attributes

+++ core-dto/src/main/java/org/yes/cart/service/dto/impl/DtoCategoryServiceImpl.java
@@ -227,8 +227,11 @@ public class DtoCategoryServiceImpl
      * {@inheritDoc}
      */
     public void remove(final long id) {
-        ((ShopCategoryService) shopCategoryGenericService).deleteAll(service.findById(id));
-        super.remove(id);
+        final Category cat = service.findById(id);
+        if (cat != null && !cat.isRoot()) {
+            ((ShopCategoryService) shopCategoryGenericService).deleteAll(service.findById(id));
+            super.remove(id);
+        }
     }
 
     /**
